requirements.txt
python-dotenv
openai
requests
pandas
streamlit